+++ run.py
@@ -544,6 +544,7 @@ def _package():
         ("Generated", "SoraConf.cs"),
         ("Generated", "SoraConfInternal.cs"),
         ("Editor", "SoraUnitySdkPostProcessor.cs"),
+        ("Editor", "SoraAndroidDependencyInjector.cs"),
     ]
     for assets_file in assets_files:
         install_file(